social_app_1/requirements.txt
# Core dependencies
streamlit>=1.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Document processing
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
PyMuPDF>=1.23.0

# LLM and Vector DB
langchain>=0.1.0
langchain-community>=0.0.15
langchain-core>=0.1.10
langsmith>=0.0.75
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Agent Orchestration
langgraph>=0.0.26
phidata>=2.0.0

# Local LLM
ollama>=0.1.5

# ML Models
scikit-learn>=1.3.2
numpy>=1.24.0

# API and UI
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.31.0
gradio>=4.13.0

# Observability
langfuse>=2.0.0